// src/enum/config.enum.ts
export enum DbEnum {
  DB_TYPE = "DB_TYPE",
  DB_HOST = "DB_HOST",
  DB_PORT = "DB_PORT",
  DB_USERNAME = "DB_USERNAME",
  DB_PASSWORD = "DB_PASSWORD",
  DB_DATABASE = "DB_DATABASE",
  DB_SYNC = "DB_SYNC",
  DB_LOGGING = "DB_LOGGING"
}

export enum JwtEnum {
  JWT_SECRET = "JWT_SECRET",
  JWT_ACCESS_TOKEN_EXPIRATION = "JWT_ACCESS_TOKEN_EXPIRATION",
  JWT_REFRESH_TOKEN_EXPIRATION = "JWT_REFRESH_TOKEN_EXPIRATION"
}